Generate int properties using the blueprint isInt flag
Upgrade @seamapi/blueprint to 0.58.0, which adds the isInt flag to
number properties, and @seamapi/types to 1.966.0, which restores the
missing created_at property on error and warning resources.

+++ src/Objects/AccessGrantRequestedAccessMethods.php
@@ -25,7 +25,7 @@ class AccessGrantRequestedAccessMethods
         public array|null $created_access_method_ids,
         public string|null $created_at,
         public string|null $display_name,
-        public float|null $instant_key_max_use_count,
+        public int|null $instant_key_max_use_count,
         public string|null $mode,
     ) {}
 }
