+++ src/Objects/DeviceActiveThermostatSchedule.php
@@ -36,7 +36,7 @@ class DeviceActiveThermostatSchedule
         public string|null $ends_at,
         public array $errors,
         public bool|null $is_override_allowed,
-        public float|null $max_override_period_minutes,
+        public int|null $max_override_period_minutes,
         public string|null $name,
         public string|null $starts_at,
         public string|null $thermostat_schedule_id,
